CAY-2659 Use new SQLBuilder utility to generate SQL for batch queries
